refactor: move geolocation button to shared ui components
