+++ .idea/workspace.xml
@@ -4,11 +4,7 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="40c482bd-ec28-48dc-8f63-37a17d928e1b" name="Changes" comment="fix: fix unused import issue">
-      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/frontend/src/components/interactive-map/filter.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/frontend/src/components/interactive-map/filter.tsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/frontend/src/pages/home/home.page.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/frontend/src/pages/home/home.page.tsx" afterDir="false" />
-    </list>
+    <list default="true" id="40c482bd-ec28-48dc-8f63-37a17d928e1b" name="Changes" comment="update: update style to fit responsiveness" />
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
@@ -97,7 +93,7 @@
       <workItem from="1722364016422" duration="8654000" />
       <workItem from="1722454505316" duration="2234000" />
       <workItem from="1722614726629" duration="3882000" />
-      <workItem from="1722855514990" duration="18816000" />
+      <workItem from="1722855514990" duration="19385000" />
     </task>
     <task id="LOCAL-00001" summary="chore: add axios">
       <option name="closed" value="true" />
@@ -243,7 +239,15 @@
       <option name="project" value="LOCAL" />
       <updated>1722874031933</updated>
     </task>
-    <option name="localTasksCounter" value="19" />
+    <task id="LOCAL-00019" summary="update: update style to fit responsiveness">
+      <option name="closed" value="true" />
+      <created>1722874368779</created>
+      <option name="number" value="00019" />
+      <option name="presentableId" value="LOCAL-00019" />
+      <option name="project" value="LOCAL" />
+      <updated>1722874368779</updated>
+    </task>
+    <option name="localTasksCounter" value="20" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -267,7 +271,8 @@
     <MESSAGE value="wip: adding filtering feature" />
     <MESSAGE value="feat: finish filter features on interactive map" />
     <MESSAGE value="fix: fix unused import issue" />
-    <option name="LAST_COMMIT_MESSAGE" value="fix: fix unused import issue" />
+    <MESSAGE value="update: update style to fit responsiveness" />
+    <option name="LAST_COMMIT_MESSAGE" value="update: update style to fit responsiveness" />
   </component>
   <component name="XDebuggerManager">
     <breakpoint-manager>
